Made a Start Scene, added some models
Parented that shit to the original placeholder, turned off Mesh Renderer.
Parented the ingredients to the bowl when the trigger is hit => destroy the ingredients  when cake is here
Needs to make trigger much smaller so they can't fuk with the bowl sides

Shit u need to do:
1. put ingredients into bowl. Look into SetActive, place the thing into the bowl already and set active the other model when ingredient is dropped.
Set inactive the child when dropped into bowl to make it disappear.
2. add instruction screen and ending screen
3. Make time increments that increase using Time.deltaTime => delay action
4. MAKE THIS LEVEL WORK PERFECTLY, THEN MAKE DUPLICATE LEVELS OF DIFFERENT BAKING ITEMS 
5. If have time, add a MIX INGREDIENT gameplay canvas that displays 2D view and changes the image from an array on mouse click