DO NOT USE INSTANTIATE.  INSTEAD USE SETACTIVE BECAUSE FUK THAT OTHER SHIT
keeps making my cake prefab disappear if I setActive(false) it. Guess because it's a prefab.
Solution: make a cake in the scene where the bowl is, put to False, then setActive(true) when all toggles are On. Later, turn it off and instantiate the Poop (the butter is the placeholder for now).
Linked all the levels in order together.